added stateful args
