+++ otito/metrics/_base_metric.py
@@ -1,8 +1,9 @@
 import inspect
-import functools
 from abc import ABC, abstractmethod
 
 from pydantic import create_model
+
+from otito.metrics.utils import validation_handler
 
 
 class BaseMetric(ABC):
@@ -10,10 +11,8 @@ class BaseMetric(ABC):
         self.callable = self.call_metric_function
         self.validate_input = kwargs["validate_input"]
         self.validator = self._build_validator(kwargs["package"], kwargs["val_config"])
-
-    @abstractmethod
-    def call_metric_function(self):
-        pass
+        self.metric_args = self.get_metric_args()
+        self.stateful = kwargs["stateful"]
 
     @abstractmethod
     def reset(self):
@@ -27,21 +26,16 @@ class BaseMetric(ABC):
     def update(self):
         pass
 
+    def get_metric_args(self):
+        arg_names = inspect.getfullargspec(self.update).args
+        arg_names.remove("self")
+        return arg_names
+
     def _build_validator(self, package, config):
         return create_model(f"{package}:{self.__class__.__name__}Model", **config)
 
-    @staticmethod
-    def validation_handler(func):
-        @functools.wraps(func)
-        def inner(self, *args, **kwargs):
-            return func(self, **self._parse_input(*args, **kwargs))
-
-        return inner
-
     def _merge_args_kwargs(self, *args, **kwargs):
-        arg_names = inspect.getfullargspec(self.callable.__wrapped__).args
-        arg_names.remove("self")
-        kwargs.update(dict(zip(arg_names, args)))
+        kwargs.update(dict(zip(self.metric_args, args)))
         return kwargs
 
     def _parse_input(self, *args, **kwargs):
@@ -51,6 +45,15 @@ class BaseMetric(ABC):
         else:
             metric_arguments = self.validator.construct(**metric_arguments).dict()
         return metric_arguments
+
+    @validation_handler
+    def call_metric_function(self, **kwargs):
+        self.update(**kwargs)
+        result = self.compute()
+
+        if not self.stateful:
+            self.reset()
+        return result
 
     def __call__(self, *args, **kwargs):
         return self.callable(**self._parse_input(*args, **kwargs))
